+++ vice/src/vdc/vdc.c
@@ -312,7 +312,7 @@ void vdc_update_memory_ptrs(unsigned int cycle)
 static void vdc_increment_memory_pointer(void)
 {
     vdc.mem_counter_inc = vdc.screen_text_cols;
-    if (vdc.raster.ycounter >= (int)vdc.raster_ycounter_max)
+    if (vdc.raster.ycounter >= vdc.raster_ycounter_max)
         vdc.mem_counter += vdc.mem_counter_inc + vdc.regs[27];
 
     vdc.raster.ycounter = (vdc.raster.ycounter + 1)
@@ -326,7 +326,7 @@ static void vdc_set_video_mode(void)
     vdc.raster.video_mode = (vdc.regs[25] & 0x80)
                             ? VDC_BITMAP_MODE : VDC_TEXT_MODE;
 
-    if (vdc.raster.ycounter > (vdc.regs[23] & 0x1f))
+    if (vdc.raster.ycounter > (unsigned int)(vdc.regs[23] & 0x1f))
         vdc.raster.video_mode = VDC_IDLE_MODE;
 }
